fetch/sec-metadata/style.https.sub.html

<!DOCTYPE html>

<link rel="author" href="mtrzos@google.com" title="Maciek Trzos">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/fetch/sec-metadata/resources/helper.js></script>
<body>

  <!-- Same-Origin request -->
  <link href="https://{{host}}:{{ports[https][0]}}/fetch/sec-metadata/resources/record-header.py?file=style-same-origin" rel="stylesheet">

  <!-- Same-Site request -->
  <link href="https://{{hosts[][www]}}:{{ports[https][0]}}/fetch/sec-metadata/resources/record-header.py?file=style-same-site" rel="stylesheet">

  <!-- Cross-Site request -->
  <link href="https://{{hosts[alt][www]}}:{{ports[https][0]}}/fetch/sec-metadata/resources/record-header.py?file=style-cross-site" rel="stylesheet">
</body>

<script>
  var same_origin_test = async_test("Same-Origin style");
  same_origin_test.step(function () {
      filename = "style-same-origin";
      expected_same_origin = 'destination="style", target="subresource", site="same-origin"';

      //  Requests from the server the saved value of the Sec-Metadata header
      same_origin_xhr = new XMLHttpRequest();
      same_origin_xhr.open("GET", "/fetch/sec-metadata/resources/get-header.py?file=" + filename);

      // Async test step triggered when the response is loaded
      same_origin_xhr.onreadystatechange = same_origin_test.step_func(function () {
        assert_header_equals(same_origin_xhr, same_origin_test, expected_same_origin)
      });
      same_origin_xhr.send();
  });

  var same_site_test = async_test("Same-Site style");
  same_site_test.step(function () {
      filename = "style-same-site";
      expected_same_site = 'destination="style", target="subresource", site="same-site"';

      //  Requests from the server the saved value of the Sec-Metadata header
      same_site_xhr = new XMLHttpRequest();
      same_site_xhr.open("GET", "/fetch/sec-metadata/resources/get-header.py?file=" + filename);

      // Async test step triggered when the response is loaded
      same_site_xhr.onreadystatechange = same_site_test.step_func(function () {
        assert_header_equals(same_site_xhr, same_site_test, expected_same_site)
      });
      same_site_xhr.send();
  });

  var cross_site_test = async_test("Cross-Site style");
  cross_site_test.step(function () {
      filename = "style-cross-site";
      expected_cross_site = 'destination="style", target="subresource", site="cross-site"';

      //  Requests from the server the saved value of the Sec-Metadata header
      cross_site_xhr = new XMLHttpRequest();
      cross_site_xhr.open("GET", "/fetch/sec-metadata/resources/get-header.py?file=" + filename);

      // Async test step triggered when the response is loaded
      cross_site_xhr.onreadystatechange = cross_site_test.step_func(function () {
        assert_header_equals(cross_site_xhr, cross_site_test, expected_cross_site)
      });
      cross_site_xhr.send();
  });
</script>
